RF > Delete duplicates #1312 (#1313)
